fix(arns): enforce undername limit
The ANTs provide the sorted array of records in resolution order. We are now enforcing the undername limits and returning a 402 to arns names that are not supported. Additional `X-ArNS-Limit` and `X-ArNS-Index` headers are added to indicate where the name/undername is in the list of records returned by the ANT.